PRG_0009 and PRG_0010
PRG_0009
Given a string and a second "word" string, we'll say that the word matches the string if it appears at the front of the string, except its first char does not need to match exactly. On a match, return the front of the string, or otherwise return the empty string. So, with the string "hippo" the word "hi" returns "hi" and "xip" returns "hip". The word will be at least length 1.

startWord("hippo", "hi") → "hi"
startWord("hippo", "xip") → "hip"
startWord("hippo", "i") → "h"

PRG_0010
Given a linked list of size N, your task is to complete the function isLengthEvenOrOdd() which contains the head of the linked list and check whether the length of linked list is even or odd.